+++ exporter/datadogexporter/internal/model/translator/metrics_translator.go
@@ -20,6 +20,7 @@ import (
 	"fmt"
 	"math"
 	"strconv"
+	"sync"
 
 	"github.com/DataDog/datadog-agent/pkg/quantile"
 	"go.opentelemetry.io/collector/pdata/pmetric"
@@ -36,6 +37,8 @@ type Translator struct {
 	prevPts *ttlCache
 	logger  *zap.Logger
 	cfg     translatorConfig
+
+	onceHostnameChanged sync.Once
 }
 
 // New creates a new translator with given options.
@@ -64,7 +67,11 @@ func New(logger *zap.Logger, options ...Option) (*Translator, error) {
 	}
 
 	cache := newTTLCache(cfg.sweepInterval, cfg.deltaTTL)
-	return &Translator{cache, logger, cfg}, nil
+	return &Translator{
+		prevPts: cache,
+		logger:  logger,
+		cfg:     cfg,
+	}, nil
 }
 
 // isCumulativeMonotonic checks if a metric is a cumulative monotonic metric
@@ -403,6 +410,24 @@ func (t *Translator) MapMetrics(ctx context.Context, md pmetric.Metrics, consume
 			host, err = t.cfg.fallbackHostnameProvider.Hostname(context.Background())
 			if err != nil {
 				return fmt.Errorf("failed to get fallback host: %w", err)
+			}
+		}
+
+		// Log related to the preview hostname feature flag.
+		// TODO (#10424): Remove this once the feature flag is enabled by default.
+		if !t.cfg.previewHostnameFromAttributes && host != "" {
+			previewHost, oldOk := attributes.HostnameFromAttributes(rm.Resource().Attributes(), true)
+			if !oldOk {
+				previewHost, _ = t.cfg.fallbackHostnameProvider.Hostname(context.Background())
+			}
+			if previewHost != host {
+				t.onceHostnameChanged.Do(func() {
+					t.logger.Warn(
+						"The hostname resolved from attributes on one of your metrics will change on a future minor version. See https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/10424",
+						zap.String("current hostname", host),
+						zap.String("future hostname", previewHost),
+					)
+				})
 			}
 		}